Moved attitude adjustment into its own flight mode
